requirements.txt
Flask
Flask-MySQL
Flask-SQLAlchemy
pymysql
boto3
botocore
mysql.connector